feat: move peer id from drpObject to hashgraph
