+++ packages/object/src/hashgraph/index.ts
@@ -1,13 +1,11 @@
-import type { Vertex_Operation as Operation, Vertex } from "@ts-drp/types";
+import { type Vertex_Operation as Operation, Vertex } from "@ts-drp/types";
 
 import { log } from "../index.js";
 import { BitSet } from "./bitset.js";
 import { linearizeMultipleSemantics } from "../linearize/multipleSemantics.js";
 import { linearizePairSemantics } from "../linearize/pairSemantics.js";
+import { computeHash } from "../utils/computeHash.js";
 import { ObjectSet } from "../utils/objectSet.js";
-
-// Reexporting the Vertex and Operation types from the protobuf file
-export type { Vertex, Operation };
 
 export type Hash = string;
 
@@ -103,6 +101,17 @@ export class HashGraph {
 		return this.resolveConflictsDRP
 			? this.resolveConflictsDRP(vertices)
 			: { action: ActionType.Nop };
+	}
+
+	createVertex(vertexOperation: Operation, vertexDependencies: Hash[]): Vertex {
+		const vertexTimestamp = Date.now();
+		return Vertex.create({
+			hash: computeHash(this.peerId, vertexOperation, vertexDependencies, vertexTimestamp),
+			peerId: this.peerId,
+			operation: vertexOperation,
+			dependencies: vertexDependencies,
+			timestamp: vertexTimestamp,
+		});
 	}
 
 	addToFrontier(vertex: Vertex) {
